JPA and hibernate
This is the first commit related to the jpa and hibernate. This project will include all the crud operation using hiberate. But this first commit only contains the add operation. before getting into the details of code. first briefly go through the concept of JPA and hibernate.

JPA stands for jakarta presistence API earlier known as  java presistance API. In a nutshell, jpa is the set of rules to standardise the imlpementation of ORM (Object-to-Rational Mapping). It has set of interfaces that need to be implement to  use. and Hibernate is of the framework that implement JPA and provide ORMs. Hibernate is the default framework for JPA implementation in Spring. Another implementation of JPA is EcpliseLink. Just mentioning it to know there are other implementations too but hiberate is most commonly use.

About this project:
 this project is the simple implementation to explain the crud using hibernate. All the work in this application is related to the terminal.
CurdDemoApplication: is the main application class. The main different thing about this calss us CommandLineRunner method.  this method is provided by spring and get a call right after all the beans get instantiated.

to communicate with database, we requires 3 things, one class which annotate as "Entity". this class uses to map the database table. In our case, this class is Student. Second, we need an interface which defines the functionalities that we can do on this table. and this interface here is StudentDao. And third class is needed to implement the the functionalities defined in Dao class. i.e. StudentDaoImpl. Impl class has a field of EntityManager. And all the database related functionalities can be performed using this entityManger