Springboot CRUD II: crud using JPA and hibernate for read
In this commit, we are targeting the The 'R' of CRUD. which stands for read.

So I am start my explaining from Student Entity and will move my way to main application class. In this commit. We added a default constructor for Student class which was missing earlier in C i.e. create. When we start fetching the data from database. it creates the entity object first. And for that it requires a default constructor (i.e. constructor without parameter). In java, if we dont create any constructor, then it assumes a default constructor without parameters. You would have no problem creating an object like this:
Student std1 = new Student();
But if you override the constructor, then you would also need  to right the default constructor if you need it somewhere or else above mentioned line would create a compilation error. So this is why we added the constructor here.

StudentDao: we are adding 3 methods related to fetching the data from database. and method names pretty much self explanatory here.

StudentDaoImpl: this is where the trick lies. So for finding the record via id, we find just use the find  method from the entityManager and it ll take the the primary key value and try to locate the data basis on that. but the limitation here is that it ll only be able to search the record base of the primary key. So what if you want to search the data on base of someother parameter. Here Query comes to aid our requirements.  and findAll and findByLastName using createQuery method of entityManager to search the data.

And then in CrudDemoApplication, we wrote the contoller methods to use the methods defined in StudentDao class