Springboot CRUD III: crud using JPA and hibernate for update
In this commit, we are targeting the The 'U' of CRUD. which stands for Update.

Most of the time, when we are updating something in the real world application, it is for the specific record. but there are some scenarios in which we requires to update multiple records at a time. Luckily for us, spring provide the support of both types for update. to demostrate the concept, we created two methods .i.e. "update" to update a single record and "multipleEmailUpdateByQuery".

StudentDaoImpl: wo single record update we use merge method provided by entityManager. If  this object is not already in the db, it will insert this record or else it will update the existing record.

For the multiple update, we use query as per demostrated in the multipleEmailUpdateByQuery method.