Springboot CRUD IV: crud using JPA and hibernate
In this commit, we are targeting the The 'D' of CRUD. which stands for Delete. this is the last letter so this pretty much sums up our crud project.

its implementation is almost identical with update. For details you can see deleteAll and DeleteRecord Method. But while developing this feature I encounter a very interesting issue. and it was quite important so I add it in the imple class as a comment. Please make sure to go through it I am sure it will also help you alot in your learning